Staged classification
This was tricky to get right, but a useful demonstration of how `classify` can
be used, and one that required a small but important change to `Classifier`:
when we classify, say, a list, we not only give you a classifier for the
elements of the list, we also provide the value we _used_ to classify those
values. This is evidence that the list is indeed not empty, which is necessary
for some client code.

We do not provide any evidence when the list is _empty_; instead, the
`ConcreteClassifier` in the tests use the same empty/non-empty distinction that
the `Classifier` does, so that we can check that they are indeed equal.