Do not lift constructor names to the type level
Suppose we have a user-defined type such as

```haskell
data X = A | B
```

Prior to this commit, we would infer something like

```haskell
UserDefined ('Constr "pkg" "modl" "A")
```

or

```
UserDefined ('Constr "pkg" "modl" "B")
```

depending on the constructor that's being used. However, this means that values
of the _same_ actual type (`X`) are now assigned _different_ inferred types
(`UserDefined ... "A"` versus `UserDefined ... "B"`). This doesn't work: for
example, when we infer a type for a list, we assume all elements in the list
must have the same type, and indeed they must; but if we infer different types
for them, things break.

Unfortunately, this means we get less type safety in some places
(`unsafeCoerceUserDefined` is now gone entirely).

Closes #6.